Move types to signer utils
